--Fixed fuckloads of warnings :p
git-svn-id: file:///home/skmp/Desktop/public_svn/nullDC@282 d5fdb02f-ce0d-0410-91b1-fe443107cc8c

+++ trunk/nulldc/nullDC/dc/gdrom/zGDROM.cpp
@@ -277,7 +277,7 @@ void WriteMem_gdrom(u32 Addr, u32 data, u32 sz)
 
 			// Ok this is crash worthy ;)
 
-			gdRM.Words[gdSR.ByteCount/2] = data;
+			gdRM.Words[gdSR.ByteCount/2] = (u16)data;
 
 			if(0 >= (gdSR.ByteCount -= 2))
 			{
